EPMLABSBRN-350 move ExerciseType and WordType enums to `model` package <eom>
