+++ src/main/kotlin/com/epam/brn/dto/TaskDtoForSentence.kt
@@ -1,7 +1,7 @@
 package com.epam.brn.dto
 
-import com.epam.brn.constant.ExerciseType
-import com.epam.brn.constant.WordType
+import com.epam.brn.model.ExerciseType
+import com.epam.brn.model.WordType
 import com.fasterxml.jackson.annotation.JsonIgnore
 import java.util.LinkedList
