+++ src/main/kotlin/com/epam/brn/upload/csv/series2/SeriesTwoRecordProcessor.kt
@@ -1,11 +1,11 @@
 package com.epam.brn.upload.csv.series2
 
-import com.epam.brn.constant.ExerciseType
-import com.epam.brn.constant.WordType
 import com.epam.brn.model.Exercise
+import com.epam.brn.model.ExerciseType
 import com.epam.brn.model.Resource
 import com.epam.brn.model.Series
 import com.epam.brn.model.Task
+import com.epam.brn.model.WordType
 import com.epam.brn.repo.ExerciseRepository
 import com.epam.brn.repo.ResourceRepository
 import com.epam.brn.repo.SeriesRepository
